+++ lib/open_weather/models/one_call/daily_weather.rb
@@ -12,8 +12,8 @@ module OpenWeather
         property 'pressure' # atmospheric pressure on the sea level, hPa
         property 'humidity' # humidity, %
         temperature_property 'dew_point' # atmospheric temperature (varying according to pressure and humidity) below which water droplets begin to condense and dew can form
-        property 'wind_speed' # wind speed
-        property 'wind_gust' # wind gust
+        speed_property 'wind_speed' # wind speed
+        speed_property 'wind_gust' # wind gust
         property 'wind_deg' # wind direction, degrees (meteorological)
         property 'clouds' # cloudiness, %
         property 'uvi' # UV index