+++ lib/open_weather/models/one_call/hourly_weather.rb
@@ -12,8 +12,8 @@ module OpenWeather
         temperature_property 'dew_point' # atmospheric temperature (varying according to pressure and humidity) below which water droplets begin to condense and dew can form
         property 'clouds' # cloudiness, %
         property 'visibility' # average visibility, meters
-        property 'wind_speed' # wind speed.
-        property 'wind_gust' # wind gust.
+        speed_property 'wind_speed' # wind speed.
+        speed_property 'wind_gust' # wind gust.
         property 'wind_deg' # wind direction, degrees (meteorological)
         property 'rain'
         property 'snow'